1(a)/requirements.txt
PyMuPDF
pytesseract
pdf2image
scikit-learn
numpy
Pillow
